requirements.txt
asgiref==3.3.3
cffi==1.14.5
cryptography==3.4.7
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
Markdown==3.3.4
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
pytz==2021.1
sqlparse==0.4.1
